requirements.txt
streamlit>=1.35.0
duckdb>=1.0.0
pandas>=2.2.2
numpy>=2.0.0
plotly>=5.22.0
anthropic>=0.28.0
python-dotenv>=1.0.1
faker>=25.2.0
pytest>=8.2.1
pyarrow>=18.0.0
